fix(save): calculate the version to save for new registry deps
