+++ lib/install/deps.js
@@ -256,11 +256,12 @@ exports.loadRequestedDeps = function (args, tree, saveToDependencies, log, next)
         child.isGlobal = true
       }
       var childName = moduleName(child)
+      child.saveSpec = computeVersionSpec(child)
       if (saveToDependencies) {
-        tree.package[saveToDependencies][childName] = child.package._requested.saveSpec
+        tree.package[saveToDependencies][childName] = child.saveSpec
       }
       if (saveToDependencies === 'optionalDependencies') {
-        tree.package.dependencies[childName] = child.package._requested.saveSpec
+        tree.package.dependencies[childName] = child.saveSpec
       }
       child.userRequired = true
       child.save = saveToDependencies
@@ -274,6 +275,23 @@ exports.loadRequestedDeps = function (args, tree, saveToDependencies, log, next)
       })
     }))
   }, andForEachChild(loadDeps, andFinishTracker(log, next)))
+}
+
+function computeVersionSpec (child) {
+  validate('O', arguments)
+  var requested = child.package._requested
+  if (requested.registry) {
+    var version = child.package.version
+    var rangeDescriptor = ''
+    if (semver.valid(version, true) &&
+        semver.gte(version, '0.1.0', true) &&
+        !npm.config.get('save-exact')) {
+      rangeDescriptor = npm.config.get('save-prefix')
+    }
+    return rangeDescriptor + version
+  } else {
+    return requested.saveSpec
+  }
 }
 
 function moduleNameMatches (name) {
