+++ lib/install/node.js
@@ -19,6 +19,8 @@ var defaultTemplate = {
   location: null,
   userRequired: false,
   existing: false,
+  save: false,
+  saveSpec: null,
   isTop: false
 }
