+++ lib/install/save.js
@@ -3,7 +3,6 @@ var fs = require('graceful-fs')
 var path = require('path')
 var writeFileAtomic = require('write-file-atomic')
 var log = require('npmlog')
-var semver = require('semver')
 var iferr = require('iferr')
 var validate = require('aproba')
 var without = require('lodash.without')
@@ -139,23 +138,6 @@ var getSaveType = exports.getSaveType = function (args) {
   else return 'dependencies'
 }
 
-function computeVersionSpec (child) {
-  validate('O', arguments)
-  var requested = child.package._requested
-  if (requested.registry) {
-    var version = child.package.version
-    var rangeDescriptor = ''
-    if (semver.valid(version, true) &&
-        semver.gte(version, '0.1.0', true) &&
-        !npm.config.get('save-exact')) {
-      rangeDescriptor = npm.config.get('save-prefix')
-    }
-    return rangeDescriptor + version
-  } else {
-    return requested.saveSpec
-  }
-}
-
 function getThingsToSave (tree) {
   validate('O', arguments)
   var toSave = tree.children.filter(function (child) {
@@ -163,7 +145,7 @@ function getThingsToSave (tree) {
   }).map(function (child) {
     return {
       name: moduleName(child),
-      spec: computeVersionSpec(child),
+      spec: child.saveSpec,
       save: child.save
     }
   })
